backend-authenticator
Moved everything from authentication.js to server.js
