+++ app/server.js
@@ -1,6 +1,7 @@
 const express = require("express");
 const { Pool } = require("pg");
 const env = require("../env.json");
+const crypto = require("crypto");
 
 const app = express();
 const hostname = "localhost";
@@ -16,6 +17,32 @@ app.use(express.static("public"))
 
 
 // TODO: everything
+const tokenStorage = {};
+
+const authentication = (req, res, next) => {
+    const authHeader = req.headers['authorization'];
+    let token = authHeader && authHeader.split(' ')[1];
+
+    if (!token || !tokenStorage[token]) {
+        return res.status(401).json({ error: "Token required."});
+    }
+    next();
+};
+
+const generateRandomToken = () => {
+    return crypto.randomBytes(32).toString("hex");
+}
+
+app.get("/api/auth/demoKey", (req, res) => {
+    const token = generateRandomToken();
+    tokenStorage[token] = true;
+    console.log(tokenStorage); //keep track of tokens created
+    return res.json({ token });
+})
+
+app.get("/pos/", authentication, (req, res) => {
+    return res.status(200).json({ message: "Token valid"});
+});
 
 // request header to return all the items in the database, with optional query to return based on category
 app.get("/api/items", (req, res) => {
